+++ src/api/index.js
@@ -3,5 +3,8 @@ import FriendsAPI from './Friends'
 import MatchesAPI from './Matches'
 import MatchmakingAPI from './Matchmaking'
 import NotificationsAPI from './Notifications'
+import ProfilesAPI from './Profiles'
 
-export { NotificationsAPI, AccountsAPI, MatchmakingAPI, FriendsAPI, MatchesAPI }
+export { NotificationsAPI, AccountsAPI, MatchmakingAPI, FriendsAPI, MatchesAPI, ProfilesAPI }
+
+