+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adicionado interface `MatchAPI` para controlar as requisições de partidas [#206](https://github.com/3C-gg/reload-frontend/issues/206).
+- Criado interface `ProfilesAPI`.
+- Adicionado tamanho `xl` de `80px` para o componente de `Avatar` no tema do chakra.
+- Criado layout `ProfileLayout` para a página de perfil [#133](https://github.com/3C-gg/reload-frontend/issues/133).
 - Adicionado método de `list` na interface de AccountsAPI.
 - Criado componente `MatchHistoryPagination`.
 - Criado componente `MatchHistoryPaginationItem`.
@@ -37,6 +40,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Alterado `MatchView` para usar a interface MatchesApi para fazer requisições de partidas.
+- Mudado extensao da interface `MatchmakingAPI` de `jsx` para `js`.
+- Atualizado componente `MatchHistoryStatsLink` para receber prop user_id.
+- Adicionado Skeleton no componente `MatchHistoryList`.
+- Atualizado componentes `SideBar` e `SideBarHeader` para para o user id para a página de perfil. 
+- Atualizado componentes `LevelStatsCard` e `HeatmapStatsCard` para ficar igual ao design.
+- Atualizado componente `HeaderProfile` para ficar igual ao design.
 - Alterado o componente `MatchHistoryStatsAccordion` para ficar igual ao design e agora é um Link que leva para a página de detalhes da partida [#212](https://github.com/3C-gg/reload-frontend/issues/212).
 - `NotificationList` passa a ir na API para popular as notificações [#164](https://github.com/3C-gg/reload-frontend/issues/164).
 - Layout `MainLayout` alterado para comportar nova sidebar [#215](https://github.com/3C-gg/reload-frontend/issues/215).
@@ -66,6 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Componente `MatchHistoryStatsLink` renderizando `NaN` quando os stats vêm zerados.
 - Componente `NotificationList` agora aparece por cima dos outros elementos da página de maneira correta [#205](https://github.com/3C-gg/reload-frontend/issues/205).
 
 ## [611cea3 - 8/5/2023]